+++ Source/CesiumRuntime/Private/Cesium3DTileset.cpp
@@ -99,7 +99,7 @@ ACesium3DTileset::ACesium3DTileset()
       _beforeMoviePreloadSiblings{PreloadSiblings},
       _beforeMovieLoadingDescendantLimit{LoadingDescendantLimit},
       _beforeMovieKeepWorldOriginNearCamera{true},
-      _tilesToNoLongerRenderNextFrame{} {
+      _beforeMovieUseLodTransitions{true} {
 
   PrimaryActorTick.bCanEverTick = true;
   PrimaryActorTick.TickGroup = ETickingGroup::TG_PostUpdateWork;
@@ -364,6 +364,7 @@ void ACesium3DTileset::PlayMovieSequencer() {
   this->_beforeMoviePreloadAncestors = this->PreloadAncestors;
   this->_beforeMoviePreloadSiblings = this->PreloadSiblings;
   this->_beforeMovieLoadingDescendantLimit = this->LoadingDescendantLimit;
+  this->_beforeMovieUseLodTransitions = this->UseLodTransitions;
   if (IsValid(this->ResolveGeoreference())) {
     this->_beforeMovieKeepWorldOriginNearCamera =
         this->ResolveGeoreference()->KeepWorldOriginNearCamera;
@@ -373,6 +374,7 @@ void ACesium3DTileset::PlayMovieSequencer() {
   this->PreloadAncestors = false;
   this->PreloadSiblings = false;
   this->LoadingDescendantLimit = 10000;
+  this->UseLodTransitions = false;
   if (IsValid(this->ResolveGeoreference())) {
     this->ResolveGeoreference()->KeepWorldOriginNearCamera = false;
   }
@@ -383,6 +385,7 @@ void ACesium3DTileset::StopMovieSequencer() {
   this->PreloadAncestors = this->_beforeMoviePreloadAncestors;
   this->PreloadSiblings = this->_beforeMoviePreloadSiblings;
   this->LoadingDescendantLimit = this->_beforeMovieLoadingDescendantLimit;
+  this->UseLodTransitions = this->_beforeMovieUseLodTransitions;
   if (IsValid(this->ResolveGeoreference())) {
     this->ResolveGeoreference()->KeepWorldOriginNearCamera =
         this->_beforeMovieKeepWorldOriginNearCamera;
@@ -1567,15 +1570,13 @@ void removeVisibleTilesFromList(
  *
  * @param tiles The tiles to hide
  */
-void hideTilesToNoLongerRender(
-    const std::vector<Cesium3DTilesSelection::Tile*>& tiles) {
+void hideTiles(const std::vector<Cesium3DTilesSelection::Tile*>& tiles) {
   for (Cesium3DTilesSelection::Tile* pTile : tiles) {
     if (pTile->getState() != Cesium3DTilesSelection::TileLoadState::Done) {
       continue;
     }
 
     const Cesium3DTilesSelection::TileContent& content = pTile->getContent();
-
     const Cesium3DTilesSelection::TileRenderContent* pRenderContent =
         content.getRenderContent();
     if (!pRenderContent) {
@@ -1586,13 +1587,39 @@ void hideTilesToNoLongerRender(
         pRenderContent->getRenderResources());
     if (Gltf && Gltf->IsVisible()) {
       Gltf->SetVisibility(false, true);
-      Gltf->SetCollisionEnabled(ECollisionEnabled::NoCollision);
     } else {
       // TODO: why is this happening?
       UE_LOG(
           LogCesium,
           Verbose,
           TEXT("Tile to no longer render does not have a visible Gltf"));
+    }
+  }
+}
+
+/**
+ * @brief Removes collision for tiles that have been removed from the render
+ * list. This includes tiles that are fading out.
+ */
+void removeCollisionForTiles(
+    const std::unordered_set<Cesium3DTilesSelection::Tile*>& tiles) {
+
+  for (Cesium3DTilesSelection::Tile* pTile : tiles) {
+    if (pTile->getState() != Cesium3DTilesSelection::TileLoadState::Done) {
+      continue;
+    }
+
+    const Cesium3DTilesSelection::TileContent& content = pTile->getContent();
+    const Cesium3DTilesSelection::TileRenderContent* pRenderContent =
+        content.getRenderContent();
+    if (!pRenderContent) {
+      continue;
+    }
+
+    UCesiumGltfComponent* Gltf = static_cast<UCesiumGltfComponent*>(
+        pRenderContent->getRenderResources());
+    if (Gltf) {
+      Gltf->SetCollisionEnabled(ECollisionEnabled::NoCollision);
     }
   }
 }
@@ -1647,6 +1674,9 @@ void ACesium3DTileset::updateTilesetOptionsFromProperties() {
   options.enforceCulledScreenSpaceError = this->EnforceCulledScreenSpaceError;
   options.culledScreenSpaceError =
       static_cast<double>(this->CulledScreenSpaceError);
+  options.enableLodTransitionPeriod = this->UseLodTransitions;
+  options.lodTransitionLength = this->LodTransitionLength;
+  // options.kickDescendantsWhileFadingIn = false;
 }
 
 void ACesium3DTileset::updateLastViewUpdateResultState(
@@ -1765,9 +1795,43 @@ void ACesium3DTileset::showTilesToRender(
 
     if (!Gltf->IsVisible()) {
       Gltf->SetVisibility(true, true);
-      Gltf->SetCollisionEnabled(ECollisionEnabled::QueryAndPhysics);
     }
+
+    Gltf->SetCollisionEnabled(ECollisionEnabled::QueryAndPhysics);
   }
+}
+
+static void updateTileFade(Cesium3DTilesSelection::Tile* pTile) {
+  if (!pTile || !pTile->getContent().isRenderContent()) {
+    return;
+  }
+
+  if (pTile->getState() != Cesium3DTilesSelection::TileLoadState::Done) {
+    return;
+  }
+
+  const Cesium3DTilesSelection::TileContent& content = pTile->getContent();
+  const Cesium3DTilesSelection::TileRenderContent* pRenderContent =
+      content.getRenderContent();
+  if (!pRenderContent) {
+    return;
+  }
+
+  UCesiumGltfComponent* pGltf = reinterpret_cast<UCesiumGltfComponent*>(
+      pRenderContent->getRenderResources());
+  if (!pGltf) {
+    return;
+  }
+
+  // Remap the fade percentage so that [0,0.5] --> [0,1] and [0.5,1] --> [1].
+  // This forces fading out tiles to stay opaque until the new tiles have fully
+  // faded in. This is needed for dithered fading, otherwise you will be able
+  // to partially see straight through both tiles during a transition.
+  float percentage =
+      pTile->getContent().getRenderContent()->getLodTransitionFadePercentage();
+  percentage = glm::clamp(2.0f * percentage, 0.0f, 1.0f);
+
+  pGltf->UpdateFade(percentage);
 }
 
 // Called every frame
@@ -1826,17 +1890,39 @@ void ACesium3DTileset::Tick(float DeltaTime) {
   }
 
   const Cesium3DTilesSelection::ViewUpdateResult& result =
-      this->_captureMovieMode ? this->_pTileset->updateViewOffline(frustums)
-                              : this->_pTileset->updateView(frustums);
+      this->_captureMovieMode
+          ? this->_pTileset->updateViewOffline(frustums)
+          : this->_pTileset->updateView(frustums, DeltaTime);
   updateLastViewUpdateResultState(result);
   this->UpdateLoadStatus();
 
+  removeCollisionForTiles(result.tilesFadingOut);
+
   removeVisibleTilesFromList(
-      this->_tilesToNoLongerRenderNextFrame,
+      _tilesToHideNextFrame,
       result.tilesToRenderThisFrame);
-  hideTilesToNoLongerRender(this->_tilesToNoLongerRenderNextFrame);
-  this->_tilesToNoLongerRenderNextFrame = result.tilesToNoLongerRenderThisFrame;
+  hideTiles(_tilesToHideNextFrame);
+
+  _tilesToHideNextFrame.clear();
+  for (Cesium3DTilesSelection::Tile* pTile : result.tilesFadingOut) {
+    Cesium3DTilesSelection::TileRenderContent* pRenderContent =
+        pTile->getContent().getRenderContent();
+    if (!this->UseLodTransitions ||
+        (pRenderContent &&
+         pRenderContent->getLodTransitionFadePercentage() <= 0.0f)) {
+      _tilesToHideNextFrame.push_back(pTile);
+    }
+  }
+
   showTilesToRender(result.tilesToRenderThisFrame);
+
+  for (Cesium3DTilesSelection::Tile* pTile : result.tilesToRenderThisFrame) {
+    updateTileFade(pTile);
+  }
+
+  for (Cesium3DTilesSelection::Tile* pTile : result.tilesFadingOut) {
+    updateTileFade(pTile);
+  }
 }
 
 void ACesium3DTileset::EndPlay(const EEndPlayReason::Type EndPlayReason) {